requirements.txt
flake8==7.1.1
flake8-docstrings==1.7.0
pyTelegramBotAPI==4.22.1
pytest==8.3.3
pytest-timeout==2.3.1
python-dotenv==1.0.1
requests==2.32.3